+++ R/import.R
@@ -93,9 +93,7 @@
 #' @importFrom tibble as_tibble is_tibble
 #' @export
 import <- function(file, format, setclass, which, ...) {
-    if (isFALSE(inherits(file, "character")) || isFALSE(length(file) == 1)) {
-        stop("Invalid `file` argument.", call. = FALSE)
-    }
+    .check_file(file, single_only = TRUE)
     if (grepl("^http.*://", file)) {
         file <- remote_to_local(file, format = format)
     }